test: keep secret permission checks Unix-only
